Conflicts:
	README
	README.install.bytecode
	README.install.source
	bin/runAllProofs.pl
	bin/runProver
	system/resources/META-INF/services/de.uka.ilkd.key.gui.macros.ProofMacro
	system/resources/de/uka/ilkd/key/proof/rules/ruleSetsDeclarations.key
	system/resources/de/uka/ilkd/key/proof/rules/seqCoreRules.key
	system/resources/de/uka/ilkd/key/proof/rules/seqRules.key
	system/resources/de/uka/ilkd/key/proof/rules/standardRules.key
	system/src/de/uka/ilkd/key/collection/ImmutableList.java
	system/src/de/uka/ilkd/key/gui/ApplyStrategy.java
	system/src/de/uka/ilkd/key/gui/InvariantConfigurator.java
	system/src/de/uka/ilkd/key/gui/KeYFileChooser.java
	system/src/de/uka/ilkd/key/gui/MainWindow.java
	system/src/de/uka/ilkd/key/gui/ProofManagementDialog.java
	system/src/de/uka/ilkd/key/gui/WindowUserInterface.java
	system/src/de/uka/ilkd/key/gui/actions/AbandonTaskAction.java
	system/src/de/uka/ilkd/key/gui/actions/SaveFileAction.java
	system/src/de/uka/ilkd/key/gui/actions/SearchInSequentAction.java
	system/src/de/uka/ilkd/key/gui/configuration/GeneralSettings.java
	system/src/de/uka/ilkd/key/gui/macros/AbstractPropositionalExpansionMacro.java
	system/src/de/uka/ilkd/key/gui/macros/OneStepProofMacro.java
	system/src/de/uka/ilkd/key/gui/macros/ProofMacro.java
	system/src/de/uka/ilkd/key/gui/macros/StrategyProofMacro.java
	system/src/de/uka/ilkd/key/gui/macros/TryCloseMacro.java
	system/src/de/uka/ilkd/key/gui/nodeviews/CurrentGoalView.java
	system/src/de/uka/ilkd/key/gui/nodeviews/DefaultTacletMenuItem.java
	system/src/de/uka/ilkd/key/gui/nodeviews/EmptySequent.java
	system/src/de/uka/ilkd/key/gui/nodeviews/InnerNodeView.java
	system/src/de/uka/ilkd/key/gui/nodeviews/MainFrame.java
	system/src/de/uka/ilkd/key/gui/nodeviews/SequentView.java
	system/src/de/uka/ilkd/key/gui/nodeviews/SequentViewSearchBar.java
	system/src/de/uka/ilkd/key/gui/prooftree/ProofTreeSearchBar.java
	system/src/de/uka/ilkd/key/java/SchemaRecoder2KeY.java
	system/src/de/uka/ilkd/key/java/Services.java
	system/src/de/uka/ilkd/key/java/recoderext/MethodSignatureSVWrapper.java
	system/src/de/uka/ilkd/key/java/recoderext/NoState.java
	system/src/de/uka/ilkd/key/java/recoderext/TwoState.java
	system/src/de/uka/ilkd/key/java/recoderext/adt/MethodSignature.java
	system/src/de/uka/ilkd/key/java/statement/For.java
	system/src/de/uka/ilkd/key/java/statement/LoopStatement.java
	system/src/de/uka/ilkd/key/java/visitor/ProgVarReplaceVisitor.java
	system/src/de/uka/ilkd/key/logic/PosInTerm.java
	system/src/de/uka/ilkd/key/logic/TermBuilder.java
	system/src/de/uka/ilkd/key/logic/label/LabelFactory.java
	system/src/de/uka/ilkd/key/logic/op/AbstractTermTransformer.java
	system/src/de/uka/ilkd/key/logic/op/Function.java
	system/src/de/uka/ilkd/key/logic/op/LocationVariable.java
	system/src/de/uka/ilkd/key/logic/op/ProgramConstant.java
	system/src/de/uka/ilkd/key/logic/op/ProgramVariable.java
	system/src/de/uka/ilkd/key/logic/op/SchemaVariableFactory.java
	system/src/de/uka/ilkd/key/logic/op/TermLabelSV.java
	system/src/de/uka/ilkd/key/logic/sort/Sort.java
	system/src/de/uka/ilkd/key/parser/KeYParser.g
	system/src/de/uka/ilkd/key/parser/SchemaVariableModifierSet.java
	system/src/de/uka/ilkd/key/parser/lexer.g
	system/src/de/uka/ilkd/key/pp/LogicPrinter.java
	system/src/de/uka/ilkd/key/pp/Notation.java
	system/src/de/uka/ilkd/key/pp/NotationInfo.java
	system/src/de/uka/ilkd/key/proof/BuiltInRuleAppIndex.java
	system/src/de/uka/ilkd/key/proof/Goal.java
	system/src/de/uka/ilkd/key/proof/Node.java
	system/src/de/uka/ilkd/key/proof/OpReplacer.java
	system/src/de/uka/ilkd/key/proof/Proof.java
	system/src/de/uka/ilkd/key/proof/delayedcut/DelayedCutProcessor.java
	system/src/de/uka/ilkd/key/proof/init/AbstractOperationPO.java
	system/src/de/uka/ilkd/key/proof/init/AbstractPO.java
	system/src/de/uka/ilkd/key/proof/init/DependencyContractPO.java
	system/src/de/uka/ilkd/key/proof/init/FunctionalOperationContractPO.java
	system/src/de/uka/ilkd/key/proof/init/JavaProfile.java
	system/src/de/uka/ilkd/key/proof/init/KeYUserProblemFile.java
	system/src/de/uka/ilkd/key/proof/init/ProblemInitializer.java
	system/src/de/uka/ilkd/key/proof/io/DefaultProblemLoader.java
	system/src/de/uka/ilkd/key/proof/io/ProblemLoader.java
	system/src/de/uka/ilkd/key/proof/io/ProofSaver.java
	system/src/de/uka/ilkd/key/proof/mgt/RuleJustificationInfo.java
	system/src/de/uka/ilkd/key/proof/mgt/SpecificationRepository.java
	system/src/de/uka/ilkd/key/rule/AbstractContractRuleApp.java
	system/src/de/uka/ilkd/key/rule/BlockContractRule.java
	system/src/de/uka/ilkd/key/rule/LoopInvariantBuiltInRuleApp.java
	system/src/de/uka/ilkd/key/rule/Taclet.java
	system/src/de/uka/ilkd/key/rule/TacletApp.java
	system/src/de/uka/ilkd/key/rule/UseOperationContractRule.java
	system/src/de/uka/ilkd/key/rule/WhileInvariantRule.java
	system/src/de/uka/ilkd/key/rule/label/TermLabelWorkerManagement.java
	system/src/de/uka/ilkd/key/rule/metaconstruct/IntroAtPreDefsOp.java
	system/src/de/uka/ilkd/key/rule/tacletbuilder/TacletGenerator.java
	system/src/de/uka/ilkd/key/speclang/BlockContract.java
	system/src/de/uka/ilkd/key/speclang/ContractFactory.java
	system/src/de/uka/ilkd/key/speclang/FunctionalOperationContract.java
	system/src/de/uka/ilkd/key/speclang/FunctionalOperationContractImpl.java
	system/src/de/uka/ilkd/key/speclang/LoopInvariant.java
	system/src/de/uka/ilkd/key/speclang/LoopInvariantImpl.java
	system/src/de/uka/ilkd/key/speclang/QueryAxiom.java
	system/src/de/uka/ilkd/key/speclang/SLEnvInput.java
	system/src/de/uka/ilkd/key/speclang/SimpleBlockContract.java
	system/src/de/uka/ilkd/key/speclang/jml/JMLSpecExtractor.java
	system/src/de/uka/ilkd/key/speclang/jml/pretranslation/jmlprelexer.g
	system/src/de/uka/ilkd/key/speclang/jml/translation/JMLSpecFactory.java
	system/src/de/uka/ilkd/key/speclang/jml/translation/JMLTranslator.java
	system/src/de/uka/ilkd/key/speclang/jml/translation/jmlparser.g
	system/src/de/uka/ilkd/key/strategy/StrategyProperties.java
	system/src/de/uka/ilkd/key/strategy/termfeature/PrimitiveHeapTermFeature.java
	system/src/de/uka/ilkd/key/symbolic_execution/strategy/AbstractCallStackBasedStopCondition.java
	system/src/de/uka/ilkd/key/symbolic_execution/strategy/ExecutedSymbolicExecutionTreeNodesStopCondition.java
	system/src/de/uka/ilkd/key/symbolic_execution/util/SymbolicExecutionUtil.java
	system/src/de/uka/ilkd/key/ui/AbstractUserInterface.java
	system/src/de/uka/ilkd/key/ui/ConsoleUserInterface.java
	system/src/de/uka/ilkd/key/util/MiscTools.java
	system/src/de/uka/ilkd/key/util/ProofStarter.java

+++ system/src/de/uka/ilkd/key/gui/lemmatagenerator/EnvironmentCreator.java
@@ -1,13 +1,13 @@
-// This file is part of KeY - Integrated Deductive Software Design 
+// This file is part of KeY - Integrated Deductive Software Design
 //
-// Copyright (C) 2001-2011 Universitaet Karlsruhe (TH), Germany 
+// Copyright (C) 2001-2011 Universitaet Karlsruhe (TH), Germany
 //                         Universitaet Koblenz-Landau, Germany
 //                         Chalmers University of Technology, Sweden
-// Copyright (C) 2011-2013 Karlsruhe Institute of Technology, Germany 
+// Copyright (C) 2011-2014 Karlsruhe Institute of Technology, Germany
 //                         Technical University Darmstadt, Germany
 //                         Chalmers University of Technology, Sweden
 //
-// The KeY system is protected by the GNU General 
+// The KeY system is protected by the GNU General
 // Public License. See LICENSE.TXT for details.
 //
 
@@ -48,19 +48,22 @@ public class EnvironmentCreator  {
 
                 ProblemInitializer pi = new ProblemInitializer(monitor, 
                                 new Services(profile, new KeYRecoderExcHandler()),
-                                false, listener );
+                                listener );
                
                 return pi.prepare(keyFile).getProofEnv();
         }
 
-        public File createDummyKeYFile(String pathForDummyFile) throws IOException {
+        private File createDummyKeYFile(String pathForDummyFile) throws IOException {
                 File file = new File(pathForDummyFile + File.separator
                                 + "lemmataGenDummy.key");
-                file.deleteOnExit();
-                String s = "\\problem{true}";
-                FileWriter writer = new FileWriter(file);
-                writer.write(s);
-                writer.close();
+                // see bug #1426
+                // file.deleteOnExit();
+                if(!file.exists()) {
+                    String s = "\\problem{true}";
+                    FileWriter writer = new FileWriter(file);
+                    writer.write(s);
+                    writer.close();
+                }
                 return file;
         }
 }
